requirements.txt
flask==2.0.1
chatterbot==1.0.4
python-dotenv==0.19.0
googlemaps==4.5.1
nltk==3.6.2
